+++ strax/plugin.py
@@ -149,6 +149,28 @@ class Plugin:
     def __deepcopy__(self):
         return self.__copy__(_deep_copy=True)
 
+    def __getattr__(self, name):
+        """
+        allow access to config parameters as attributes
+        this allows backwards compatibility in cases where
+        a descriptor style config depends on a non descriptor
+        style config.
+        """
+        if name == 'config':
+            raise AttributeError('Plugin not configured yet.')
+        if hasattr(self, 'config') and name in self.config:
+            message = '''
+            Looks like you are mixing config paradigms,
+            this is not recommended.
+            '''
+            warn(message, UserWarning)
+            if isinstance(self.takes_config[name], strax.Config):
+                return self.takes_config[name].__get__(self)
+            return self.config[name]
+            
+
+        raise AttributeError(f'{self.__class__.__name__} instance has no attribute {name}')
+
     def fix_dtype(self):
         try:
             # Infer dtype should always precede self.dtype (e.g. due to
@@ -437,7 +459,7 @@ class Plugin:
                         f"Plugin {d} terminated without fetching last {d}!")
 
             # This can happen especially in time range selections
-            if int(self.save_when) != strax.SaveWhen.NEVER:
+            if int(self.save_when) > strax.SaveWhen.EXPLICIT:
                 for d, buffer in self.input_buffer.items():
                     # Check the input buffer is empty
                     if buffer is not None and len(buffer):
@@ -495,7 +517,7 @@ class Plugin:
 
             # For non-saving plugins, don't be strict, just take whatever
             # endtimes are available and don't check time-consistency
-            if int(self.save_when) == strax.SaveWhen.NEVER:
+            if int(self.save_when) <= strax.SaveWhen.EXPLICIT:
                 # </start>This warning/check will be deleted, see UserWarning
                 if len(set(tranges.values())) != 1:
                     end = max([v.end for v in kwargs.values()])  # Don't delete
@@ -869,7 +891,7 @@ class CutPlugin(Plugin):
 class MergeOnlyPlugin(Plugin):
     """Plugin that merges data from its dependencies
     """
-    save_when = SaveWhen.NEVER
+    save_when = SaveWhen.EXPLICIT
 
     def infer_dtype(self):
         deps_by_kind = self.dependencies_by_kind()